Support async drivers
Create a new Driver superclass. Create a subclass AsyncDriver supporting
drivers which always produce data asynchronously.
